FIX: Split utils.py into tools and utils.
